user router protection + some books routes (Close #61, Related #60)

+++ server/routers/recipeBook.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const database = require('../module/database');
 const conn = database.conn;
+const needAuth = require('../module/token').checkApiKey;
 var bodyParser = require('body-parser');
 
 router.use(bodyParser.json());
@@ -15,6 +16,13 @@ router.use(
 async function bookExist(bookId){
   let exists = await conn.query(`SELECT COUNT(id) FROM recipe_books WHERE id=?;`, [bookId]);
   return exists[0]['COUNT(id)'] <= 0;
+}
+
+async function userHasReadAccess(bookId, userId){
+  const isPublic = await conn.query(`SELECT id FROM recipe_books WHERE id = ? AND visibility = 1`, [bookId]);
+  const isOwner = await conn.query(`SELECT id FROM recipe_books WHERE id = ? AND owner = ? `, [bookId, userId]);
+  const isAuthorized = await conn.query(`SELECT id FROM recipe_books JOIN recipe_book_access ON recipe_books.id = recipe_book_access.bookId WHERE recipe_book_access.bookId = ? AND recipe_book_access.userId = ?;`, [bookId, userId]);
+  return isPublic.length >= 1 || isOwner.length >= 1 || isAuthorized.length >= 1;
 }
 
 /**
@@ -34,7 +42,7 @@ async function bookExist(bookId){
  *   "400":
  *      description: "Name parameter is wrong"
  */
-router.post("/create", async (req, res) => {
+router.post("/create", needAuth, async (req, res) => {
     if (req.query.name === undefined){
         res.status(400).send("Need to specify a book name");
         return;
@@ -43,7 +51,7 @@ router.post("/create", async (req, res) => {
     const reg = /^[a-zA-Z0-9\s]+$/; // only chars, spaces and numbers 
     if (reg.test(req.query.name) && req.query.name.length <= 32){
         try {
-            await conn.query(`INSERT INTO recipe_books VALUES (null, ?, null, 1);`, [req.query.name]);
+            await conn.query(`INSERT INTO recipe_books VALUES (null, ?, ?, 1);`, [req.query.name, req.user.userId]);
         } catch (error) {
             console.log(error)
             res.sendStatus(500);
@@ -75,14 +83,14 @@ router.post("/create", async (req, res) => {
  *   "400":
  *      description: "BookId parameter is wrong"
  */
-router.delete("/delete", async (req, res) => {
+router.delete("/delete", needAuth, async (req, res) => {
   if (req.query.bookId === undefined){
     res.status(400).send("You need to specify a bookId to delete");
     return;
   }
   try {
     // Need to check if the req is the owner
-    await conn.query(`DELETE FROM recipe_books WHERE id=?;`, [req.query.bookId]);
+    await conn.query(`DELETE FROM recipe_books WHERE id=? AND owner = ?;`, [req.query.bookId, req.user.userId]);
   } catch (error) {
     console.log(error)
     res.sendStatus(500);
@@ -115,17 +123,19 @@ router.delete("/delete", async (req, res) => {
  *   "500":
  *      description: "The book you want to share does not exists"
  */
-router.post("/share", async (req, res) => {
+router.post("/share", needAuth, async (req, res) => {
   if (req.query.bookId === undefined || req.query.userId === undefined){
     res.status(400).send("Specify a bookId and the user you want to share the book with");
     return;
   }
   try {
-    // Need to check if the req is the owner
     if (await bookExist(req.query.bookId)){
       res.status(500).send("Can't share given book: no matching id");
       return;
     }
+    const book = await conn.query(`SELECT id FROM recipe_books WHERE id = ? AND owner = ?`, [req.query.bookId, req.user.userId]);
+    if (book.length != 1) return res.sendStatus(500); // Status 500 if requester is not the owner
+    // TODO: Check for duplicates
     await conn.query(`INSERT INTO recipe_book_access VALUES (?,?);`, [req.query.bookId, req.query.userId]);
   } catch (error) {
     console.log(error)
@@ -149,12 +159,14 @@ router.post("/share", async (req, res) => {
  * responses:
  *   "200":
  *     description: "Book has been created"
+ *   "204":
+ *     description: "Problem with the read permission"
  *   "400":
  *     description: "Something wrong with bookId"
  *   "500":
  *      description: "The book does not exists"
  */
-router.get("/recipes", async (req, res) => {
+router.get("/recipes", needAuth, async (req, res) => {
   if (req.query.bookId === undefined){
       res.status(400).send("Need to specify a bookId");
       return;
@@ -164,6 +176,7 @@ router.get("/recipes", async (req, res) => {
         res.status(500).send("Can't get recipes from given book: no matching id");
         return;
       }
+      if (!(await userHasReadAccess(req.query.bookId, req.user.userId))) return res.sendStatus(204);
       const recipesRaw = await conn.query(`SELECT recipeId FROM recipe_book_links WHERE bookId = ?;`, [req.query.bookId]);
       let recipes = Array.from(recipesRaw);
       let ids = [];
@@ -195,10 +208,12 @@ router.get("/recipes", async (req, res) => {
  * responses:
  *   "200":
  *     description: "The general informations"
+ *   "204":
+ *     description: "User doesn't have access to the book"
  *   "400":
  *      description: "BookId parameter is wrong"
  */
-router.get("/general_information", async (req, res) => {
+router.get("/general_information", needAuth, async (req, res) => {
   if (req.query.bookId === undefined){
     res.status(400).send("You need to specify a bookId to search for");
     return;
@@ -209,10 +224,13 @@ router.get("/general_information", async (req, res) => {
         if (bookInfo.length <= 0){
             res.sendStatus(204); // No book found with the given id
             return;
-        // add an else if to know if we are authorized to read it
+        } else if (! (await userHasReadAccess(req.query.bookId, req.user.userId))){
+          console.log("SUER ID = " + req.user.userId);
+          return res.sendStatus(204);
         } else {
             let countRes =  await conn.query(`SELECT COUNT(bookId) FROM recipe_book_links WHERE bookId=?;`, [req.query.bookId]);
             let count = Number(countRes[0]['COUNT(bookId)']);
+            console.log("BOOK INFO = " + bookInfo);
             res.send(JSON.stringify({"id": bookInfo[0]["id"], "name":  bookInfo[0]["name"], "owner": bookInfo[0]["owner"], "visibility": bookInfo[0]["visibility"], "recipe_count": count }));
             return;
         }
